split tool reg out
