+++ integrations/openclaw/src/index.ts
@@ -3,7 +3,8 @@ import { handleRecall } from './handlers/recall.js';
 import { handleAugmentation } from './handlers/augmentation.js';
 import { OpenClawEvent, OpenClawContext, MemoriPluginConfig } from './types.js';
 import { PLUGIN_CONFIG } from './constants.js';
-import { MemoriLogger, createRecallClient } from './utils/index.js';
+import { MemoriLogger } from './utils/index.js';
+import { registerAllTools } from './tools/index.js';
 
 const memoriPlugin = {
   id: PLUGIN_CONFIG.ID,
@@ -39,115 +40,7 @@ const memoriPlugin = {
       handleAugmentation(event as OpenClawEvent, ctx as OpenClawContext, config, logger)
     );
 
-    api.registerTool({
-      name: 'memori_recall',
-      label: 'Recall Memory',
-      description:
-        'Explicitly fetch relevant memories from Memori using filters like date, project, session, signal, and source.',
-      parameters: {
-        type: 'object',
-        properties: {
-          dateStart: {
-            type: 'string',
-            description: 'ISO 8601 date string to filter memories created on or after this time',
-          },
-          dateEnd: {
-            type: 'string',
-            description: 'ISO 8601 date string to filter memories created on or before this time',
-          },
-          projectId: {
-            type: 'string',
-            description: 'Filter to a specific project. Defaults to the current project.',
-          },
-          sessionId: {
-            type: 'string',
-            description: 'Filter to a specific session. Requires projectId to also be provided.',
-          },
-          signal: {
-            type: 'string',
-            description: 'Filter to a specific fact signal (e.g., system, user, derived)',
-          },
-          source: {
-            type: 'string',
-            description: 'Filter to a specific source origin',
-          },
-        },
-      },
-      execute: async (
-        _toolCallId: string,
-        params: {
-          dateStart?: string;
-          dateEnd?: string;
-          projectId?: string;
-          sessionId?: string;
-          signal?: string;
-          source?: string;
-        }
-      ) => {
-        try {
-          console.log('memori_recall params', JSON.stringify(params, null, 2));
-          const client = createRecallClient(config.apiKey, config.entityId);
-          const result = await client.agentRecall(params);
-          return { content: [{ type: 'text', text: JSON.stringify(result) }], details: null };
-        } catch (e) {
-          logger.warn(`memori_recall failed: ${String(e)}`);
-          return {
-            content: [{ type: 'text', text: JSON.stringify({ error: 'Recall failed' }) }],
-            details: null,
-          };
-        }
-      },
-    });
-
-    api.registerTool({
-      name: 'memori_recall_summary',
-      label: 'Recall Memory Summary',
-      description:
-        'Fetch summarized views of stored memories from Memori. Useful for getting a high-level overview of what is known about a project or session within a specific date range.',
-      parameters: {
-        type: 'object',
-        properties: {
-          dateStart: {
-            type: 'string',
-            description: 'ISO 8601 date string to filter summaries created on or after this time',
-          },
-          dateEnd: {
-            type: 'string',
-            description: 'ISO 8601 date string to filter summaries created on or before this time',
-          },
-          projectId: {
-            type: 'string',
-            description: 'Filter to a specific project. Defaults to the current project.',
-          },
-          sessionId: {
-            type: 'string',
-            description: 'Filter to a specific session. Requires projectId to also be provided.',
-          },
-        },
-      },
-      execute: async (
-        _toolCallId: string,
-        params: {
-          dateStart?: string;
-          dateEnd?: string;
-          projectId?: string;
-          sessionId?: string;
-        }
-      ) => {
-        try {
-          console.log('memori_recall_summary params', JSON.stringify(params, null, 2));
-          const client = createRecallClient(config.apiKey, config.entityId);
-          const result = await client.agentRecallSummary(params);
-          return { content: [{ type: 'text', text: JSON.stringify(result) }], details: null };
-        } catch (e) {
-          logger.warn(`memori_recall_summary failed: ${String(e)}`);
-          return {
-            content: [{ type: 'text', text: JSON.stringify({ error: 'Recall summary failed' }) }],
-            details: null,
-          };
-        }
-      },
-    });
+    registerAllTools({ api, config, logger });
   },
 };
