+++ integrations/openclaw/src/tools/types.ts
@@ -1,3 +1,9 @@
+import type { OpenClawPluginApi } from 'openclaw/plugin-sdk';
+import type { MemoriPluginConfig } from '../types.js';
+import type { MemoriLogger } from '../utils/logger.js';
+
 export interface ToolDeps {
   api: OpenClawPluginApi;
+  config: MemoriPluginConfig;
+  logger: MemoriLogger;
 }
